backend/requirements.txt
firebase_admin==6.7.0
Flask==2.2.2
flask_cors==5.0.1
openai==1.72.0
Pillow==11.1.0
protobuf==6.30.2
PyJWT==2.10.1
python-dotenv==1.1.0